add room, split up landing
