+++ client/.eslintrc.json
@@ -1,16 +1,26 @@
 {
-    "extends": ["plugin:prettier/recommended", "plugin:vue/recommended", "prettier/vue", "eslint:recommended"],
-    "plugins": ["prettier"],
-    "rules": {
-        "prettier/prettier": ["error"]
-    },
-    "parserOptions": {
-        "parser": "babel-eslint",
-        "ecmaVersion": 2017,
-        "sourceType": "module"
-    },
-    "env": {
-        "node": true,
-        "browser": true
-    }
+  "extends": [
+    "plugin:prettier/recommended",
+    "plugin:vue/recommended",
+    "prettier/vue",
+    "eslint:recommended"
+  ],
+  "plugins": [
+    "prettier",
+    "vue"
+  ],
+  "rules": {
+    "prettier/prettier": [
+      "error"
+    ]
+  },
+  "parserOptions": {
+    "parser": "babel-eslint",
+    "ecmaVersion": 8,
+    "sourceType": "module"
+  },
+  "env": {
+    "browser": true,
+    "node": true
+  }
 }
