+++ client/src/components/Login.vue
@@ -1,25 +1,23 @@
 <template>
   <div class="card-container">
-    <div class="card-header">
-      <div class="card-title">
-        <h3>Sign in</h3>
-        <hr />
-      </div>
-    </div>
-    <div class="user-input">
-      <form @submit.prevent="loginUser">
-        <div class="user-name">
-          <label for="name">Username:</label>
-          <input
-            v-model="user"
-            type="text"
-            placeholder="username"
-            class="input"
-          />
-          <input type="submit" class="submit-button" value="Set" />
-        </div>
-      </form>
-    </div>
+    <form class="md-layout" @submit.prevent="loginUser">
+      <md-card class="md-layout-item md-size-50 md-small-size-100">
+        <md-card-header>
+          <div class="md-title">Log in</div>
+        </md-card-header>
+        <md-card-content>
+          <md-field>
+            <label for="name">Username:</label>
+            <md-input v-model="user" type="text" class="input" />
+          </md-field>
+          <md-card-actions>
+            <md-button type="submit" class="md-raised md-primary"
+              >Set</md-button
+            >
+          </md-card-actions>
+        </md-card-content>
+      </md-card>
+    </form>
   </div>
 </template>
 
@@ -43,12 +41,10 @@ export default {
 </script>
 <style>
 .card-container {
-  background-color: #eee;
-  width: 50%;
+  margin-left: 35%;
+  margin-top: 10%;
+  height: 40vh;
   padding: 10px;
-  border-style: line;
-  border-radius: 2px;
-  box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.2), 0 3px 10px 0;
 }
 .card-header {
   text-align: center;
