+++ client/src/router.js
@@ -1,6 +1,6 @@
 import Vue from "vue";
 import Router from "vue-router";
-import Home from "./views/Home.vue";
+import Landing from "./views/Landing";
 
 Vue.use(Router);
 
@@ -12,17 +12,11 @@ export default new Router({
     },
     {
       path: "/",
-      name: "home",
-      component: Home
-    },
-    {
-      path: "/home",
       name: "landing",
       // route level code-splitting
       // this generates a separate chunk (about.[hash].js) for this route
       // which is lazy-loaded when the route is visited.
-      component: () =>
-        import(/* webpackChunkName: "about" */ "./views/Landing.vue")
+      component: Landing
     }
   ]
 });
