+++ client/src/views/Landing.vue
@@ -7,34 +7,7 @@
         <md-input v-model="type"></md-input>
         <span class="md-helper-text">Search for a chat!</span>
       </md-field>
-      <md-button class="md-raised md-primary" @click="openForm"
-        >Create a room</md-button
-      >
-
-      <div v-if="create" class="form">
-        <md-card md-with-hover>
-          <md-card-header>
-            <div class="md-title">Create a room:</div>
-            <div class="md-subhead">choose the values below</div>
-          </md-card-header>
-          <md-card-content>
-            <md-field md-inline>
-              <label>Topic</label>
-              <md-input v-model="topic"></md-input>
-            </md-field>
-            <md-field>
-              <label>Size</label>
-              <md-input v-model="number" type="number" min="2"></md-input>
-            </md-field>
-          </md-card-content>
-
-          <md-card-actions>
-            <md-button @click="createRoom">Create</md-button>
-            <md-button @click="formClose">Close</md-button>
-          </md-card-actions>
-        </md-card>
-      </div>
-      <Chat :user="user" />
+      <Chat :user="user" :socket="socket" />
     </div>
     <Login v-else @loginUser="loginUser" />
   </div>
@@ -53,27 +26,8 @@ export default {
   },
   data() {
     return {
-      /* user: {
-          type: String,
-          default: ""
-      },
-      login: {
-          type: Boolean,
-          default: false
-      },
-      create: {
-          type: Boolean,
-          default: false
-      },
-      number:  {
-          type: Number,
-          default: 2
-      },
-      topic: {
-          type: String,
-          default: ""
-      }, */
       user: "",
+      type: "",
       login: false,
       create: false,
       number: 2,
